refactoring

+++ PushkaGraphGui/VertexSettings.cs
@@ -13,7 +13,7 @@ namespace PushkaGraphGUI
         public static readonly Brush BoundColor = Brushes.Black;
         public static readonly int BoundThickness = 2;
         public static readonly int BoundThicknessHover = 4;
-        public static readonly int Size = 32;
+        public static readonly double Size = 32;
         public static readonly int ZIndex = 1;
         public static readonly int MovingZIndex = 2;
     }
